+++ packages/ui-core/CHANGELOG.md
@@ -14,16 +14,20 @@
 - **Select**
     - fixed hover with touch events detector
     - add props label
+    - add test
 - **Textarea**
     - fixed hover with touch events detector
     - add props label
+    - add test
 - **TextField**
     - fixed hover with touch events detector
     - add props label
+    - add tests
 - **package.json**
     - updated deep-equal module
 - **InputLabel**
     - add component for input label
+    - add types and tests
 
 ## Release 1.0.12
 - **Paragraph**